int/ast/is: renames the `astw` package to simplify calls to its last remaining symbol `Nil`, which performs nil-checks on all possible `nil` values of `ast.Node`.
