+++ internal/evolution/genetics/replace.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"go/ast"
 
-	"tde/internal/ast/astwutl"
+	"tde/internal/ast/is"
 )
 
 // Replaces the (c)urrent node with (n)ext node by finding the "field" on its parent, not directly.
@@ -17,7 +17,7 @@ func replaceOnParent(root ast.Node, c, n any) bool {
 	replaced := false
 
 	ast.Inspect(root, func(m ast.Node) bool {
-		if root == m || m == nil || astwutl.IsNodeNil(m) {
+		if root == m || m == nil || is.Nil(m) {
 			return !found
 		}
